chore: 최신

+++ client/src/components/common/card/review_card/CardBack.jsx
@@ -7,6 +7,7 @@ import disLikesIcon from '../../../../assets/icons/heart.png';
 import likesIcon from '../../../../assets/icons/heart_filled.png';
 import bookApi from '../../../../api/bookApi';
 import { useSelector } from 'react-redux';
+import basicProfileImage from '../../../../assets/user.png';
 
 export default function CardBack({
   book,
@@ -51,6 +52,8 @@ export default function CardBack({
         await bookApi.postBookLikes(id);
         setBookLikeCount((prev) => prev + 1);
       } else {
+        if (bookLikeCount === 0) return;
+
         await bookApi.deleteBookLikes(id);
         setBookLikeCount((prev) => prev - 1);
       }
@@ -91,7 +94,7 @@ export default function CardBack({
       <div className={styles.bottomWrapper}>
         <div className={clsx(styles.flexBox, styles.justifyStart)} onClick={clickProfileHandler}>
           <div className={styles.profileImageWrapper}>
-            <img src={profileImageUrl} alt="profileImage" />
+            <img src={profileImageUrl || basicProfileImage} alt="profileImage" />
           </div>
           <div className={styles.nickname}>{nickname}</div>
         </div>
@@ -112,7 +115,7 @@ export default function CardBack({
             />
           )}
           <div className={styles.grayText}>좋아요</div>
-          <div className={styles.grayText}>{bookLikeCount}</div>
+          <div className={clsx(styles.grayText, styles.countWidth)}>{bookLikeCount}</div>
         </div>
       </div>
     </div>
